update governance and feSplit contract

+++ smartcontract/service/native/states/governance.go
@@ -53,7 +53,7 @@ type PeerPoolList struct {
 }
 
 type PeerPool struct {
-	Index      uint64   `json:"index"`
+	Index      uint32   `json:"index"`
 	PeerPubkey string   `json:"peerPubkey"`
 	Status     Status   `json:"status"`
 	QuitView   *big.Int `json:"quitView"`
@@ -82,7 +82,7 @@ type VoteInfoPool struct {
 }
 
 type PeerStakeInfo struct {
-	Index      uint64 `json:"index"`
+	Index      uint32 `json:"index"`
 	PeerPubkey string `json:"peerPubkey"`
 	Stake      uint64 `json:"stake"`
 }
